backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12  # Добавляем эту строку